@@ -4,18 +4,20 @@ buildscript {
     repositories {
         mavenCentral()
         jcenter()
+        google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.3.3'
+        classpath 'com.android.tools.build:gradle:4.1.1'
     }
 }
 
 // To avoid manually setting the same values in all Android modules, set the value on the root
 // project and then reference this from the modules
 ext {
-    compileSdkVersion = 25
-    targetSdkVersion = 25
-    buildToolsVersion = "25.0.3"
+    compileSdkVersion = 31
+    targetSdkVersion = 31
+    minSdkVersion = 21
+    buildToolsVersion = "31.0.0"
 }
 
 def isReleaseBuild() {
@@ -29,5 +31,6 @@ allprojects {
     repositories {
         mavenCentral()
         jcenter()
+        google()
     }
 }
