@@ -3,7 +3,7 @@ package pl.charmas.android.reactivelocation.sample;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.widget.TextView;
 
 import com.google.android.gms.location.places.Place;
